+++ apps/demo/src/components/studio/useValidator.ts
@@ -1,29 +1,27 @@
 import { useEffect, useState } from "react";
 import { Validator } from "@cfworker/json-schema";
 
-const schemaUrl =
-  "/GITHUB_CONTENT_URL/versa-protocol/schema/main/data/receipt.schema.json";
-
 type InitialValidator = null | Validator;
 
 let validator: InitialValidator = null;
 
-const getValidator = async () => {
+const getValidator = async (schemaVersion: string) => {
   if (validator) {
     return validator;
   }
+  const schemaUrl = `/GITHUB_CONTENT_URL/versa-protocol/schema/${schemaVersion}/data/receipt.schema.json`;
   const schema = await (await fetch(schemaUrl)).text();
   validator = new Validator(JSON.parse(schema));
   return validator;
 };
 
-export const useValidator = () => {
+export const useValidator = (schemaVersion: string) => {
   const [validator, setValidator] = useState<InitialValidator>();
   useEffect(() => {
     (async function () {
-      const v = await getValidator();
+      const v = await getValidator(schemaVersion);
       setValidator(v);
     })();
-  }, [setValidator]);
+  }, [setValidator, schemaVersion]);
   return { validator };
 };