+++ packages/belt/src/lts/1.8.0/receiptHelpers.test.ts
@@ -1,7 +1,7 @@
-import { Receipt } from "@versaprotocol/schema";
+import { lts } from "@versaprotocol/schema";
 import {
   aggregateTaxes,
-  determineTicketFare,
+  aggregateTicketFares,
   organizeFlightTickets,
   organizeSegmentedItineraries,
   organizeTransitRoutes,
@@ -9,7 +9,7 @@ import {
 
 import { receipts } from "@versaprotocol/examples";
 
-const railReceipt = receipts.rail;
+const railReceipt = receipts.rail as unknown as lts.v1_8_0.Receipt;
 
 const testData = {
   general: null,
@@ -21,13 +21,9 @@ const testData = {
   flight: {
     tickets: [
       {
-        fare: 123, // silly data, but an important edgecase to test
-        taxes: [],
-        metadata: [],
         segments: [
           {
             fare: 15233,
-            metadata: [],
             departure_airport_code: "SEA",
             arrival_airport_code: "MSP",
             departure_at: 1713196492,
@@ -56,11 +52,10 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
           {
             fare: 12186,
-            metadata: [],
             departure_airport_code: "MSP",
             arrival_airport_code: "GFK",
             departure_at: 1713196492,
@@ -89,11 +84,10 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
           {
             fare: 14014,
-            metadata: [],
             departure_airport_code: "GFK",
             arrival_airport_code: "MSP",
             departure_at: 1713196492,
@@ -122,11 +116,10 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
           {
             fare: 19498,
-            metadata: [],
             departure_airport_code: "MSP",
             arrival_airport_code: "SEA",
             departure_at: 1713196492,
@@ -155,7 +148,7 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
         ],
         number: "0062698215636",
@@ -163,12 +156,9 @@ const testData = {
         passenger: "Jimmy Dean",
       },
       {
-        taxes: [],
-        metadata: [],
         segments: [
           {
             fare: 15233,
-            metadata: [],
             departure_airport_code: "SEA",
             arrival_airport_code: "MSP",
             departure_at: 1713196492,
@@ -197,11 +187,10 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
           {
             fare: 12186,
-            metadata: [],
             departure_airport_code: "MSP",
             arrival_airport_code: "GFK",
             departure_at: 1713196492,
@@ -230,11 +219,10 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
           {
             fare: 14014,
-            metadata: [],
             departure_airport_code: "GFK",
             arrival_airport_code: "MSP",
             departure_at: 1713196492,
@@ -263,11 +251,10 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
           {
             fare: 19498,
-            metadata: [],
             departure_airport_code: "MSP",
             arrival_airport_code: "SEA",
             departure_at: 1713196492,
@@ -296,7 +283,7 @@ const testData = {
                 name: "US Flight Segment Tax",
               },
             ],
-            adjustments: [],
+            adjustments: null,
           },
         ],
         number: "0062698215637",
@@ -305,7 +292,7 @@ const testData = {
       },
     ],
     itinerary_locator: "1122337694093",
-    invoice_level_adjustments: [],
+    invoice_level_adjustments: null,
   },
 };
 
@@ -319,14 +306,12 @@ describe("aggregateTaxes", () => {
   });
 });
 
-describe("determineTicketFare", () => {
-  const organizedTickets = organizeFlightTickets(testData.flight); // TODO
+describe("aggregateTicketFares", () => {
+  const organizedTickets = organizeFlightTickets(testData.flight as any); // TODO
   // This is sort of just a reducer but that's fine
   it("should work", () => {
-    let result = determineTicketFare(testData.flight.tickets[1]);
+    let result = aggregateTicketFares(organizedTickets[0]);
     expect(result).toBe(60931);
-    expect(organizedTickets[0].passengers[0].fare).toBe(123);
-    expect(organizedTickets[0].passengers[1].fare).toBe(60931);
   });
 });
 
@@ -397,7 +382,7 @@ describe("organizeTransitRoutes", () => {
 
 describe("organizeSegmentedItineraries", () => {
   it("should organized itineraries into two grouped itineraries for an outgoing itinerary and a return", () => {
-    const receipt = receipts.flight as Receipt;
+    const receipt = receipts.flight as unknown as lts.v1_8_0.Receipt;
 
     let ticket = receipt?.itemization?.flight?.tickets[0];
     if (!ticket || !ticket.segments) {
@@ -414,17 +399,5 @@ describe("organizeSegmentedItineraries", () => {
     expect(organized[1].departure_at).toBe(1713900952);
     expect(organized[1].departure_city).toBe("Grand Forks");
     expect(organized[1].arrival_city).toBe("Seattle");
-  });
-});
-
-describe("determineTicketFare", () => {
-  // This is sort of just a reducer but that's fine
-  it("should aggregate segment level data", () => {
-    let result = determineTicketFare(testData.flight.tickets[1]);
-    expect(result).toBe(60931);
-  });
-  it("should use ticket fare if populated", () => {
-    let result = determineTicketFare(testData.flight.tickets[0]);
-    expect(result).toBe(123);
   });
 });