+++ packages/react/release.config.ts
@@ -14,11 +14,6 @@ const config: RollupOptions = {
   output: [
     {
       dir: "dist",
-      format: "cjs",
-      sourcemap: true,
-    },
-    {
-      dir: "dist",
       format: "es",
       exports: "named",
       sourcemap: true,