+++ apps/storybook/CHANGELOG.md
@@ -1,11 +1,19 @@
 # versa-storybook
 
-## 0.0.32-alpha.0
+## 0.0.33
 
 ### Patch Changes
 
 - Updated dependencies
-  - @versaprotocol/react@1.8.8-alpha.0
+  - @versaprotocol/react@1.9.0
+
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/examples@1.9.0
+  - @versaprotocol/react@1.8.8
 
 ## 0.0.31