+++ packages/pdfgen/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @versaprotocol/pdfgen
 
+## 1.9.0
+
+### Minor Changes
+
+- Update to schema version 1.9, supporting taxes and fare on the flight ticket level
+
+### Patch Changes
+
+- Updated dependencies
+  - @versaprotocol/schema@1.9.0
+  - @versaprotocol/belt@1.9.0
+
 ## 1.8.9
 
 ### Patch Changes