+++ packages/react/CHANGELOG.md
@@ -1,10 +1,19 @@
 # @versaprotocol/react
 
-## 1.8.8-alpha.0
+## 1.9.0
+
+### Minor Changes
+
+- Update to schema version 1.9, fix fare display bug
+
+## 1.8.8
 
 ### Patch Changes
 
-- Update to React 19
+- Updated dependencies
+  - @versaprotocol/schema@1.9.0
+  - @versaprotocol/belt@1.9.0
+  - @versaprotocol/pdfgen@1.9.0
 
 ## 1.8.7